mm, sparse: remove check with __highest_present_section_nr in for_each_present_section_nr()
A valid present section number is in [0, __highest_present_section_nr].
And the return value of next_present_section_nr() meets this
requirement. This means it is not necessary to check it with
__highest_present_section_nr again in for_each_present_section_nr().

Since we pass an unsigned long *section_nr* to
for_each_present_section_nr(), we need to cast it to int before
comparing.

Signed-off-by: Wei Yang <richard.weiyang@gmail.com>