btrfs: improve error handling of btrfs_add_link()
err holds the return value of either btrfs_del_root_ref() or
btrfs_del_inode_ref() but it hasn't been checked since it's
introduction with commit fe66a05a0679 (Btrfs: improve error handling
for btrfs_insert_dir_item callers) in 2012.

To quote David:

"If the error handling in the error handling fails, there's not much left to do
and the abort either happened earlier in the callees or is necessary here."

So if one of btrfs_del_root_ref() or btrfs_del_inode_ref() failed, abort the
transaction.

Link: https://lore.kernel.org/linux-btrfs/20181126183759.GJ2842@twin.jikos.cz/
Signed-off-by: Johannes Thumshirn <jthumshirn@suse.de>