Add Red Hat's ibdev2netdev helper
This is a helper script that Red Hat had previously been including in it's
ibutils package, which is no more, with the retirement of ibutils and
ibutils2 being proprietary to Mellanox now. This script still has use, and
needs a better home, so we'd like to add it to rdma-core's redhat/
directory -- if not somewhere more generic, should other distros wish to
make use of it as well.

CC: Doug Ledford <dledford@redhat.com>
CC: Honggang Li <honli@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>