ALSA: hda/realtek - Add quirks for Dell Optiplex 5060, 7060 for audio out while using headphones
There are two audio ports on these machines - headset and mic.
There' s no sound when hp is inserted into the headset port,
though the hp is detected. Without pulseaudio, this issue is
seen on both the Optiplex machines. The issue goes away with the
"options snd-hda-intel model=dell-headset-dock" quirk in the
/etc/modprobe.d/alsa-base.conf or /etc/modprobe.d/alsa.conf
Getting this fix into the kernel instead of fixing in an
application layer above, or in a .conf file. This is also helpful
for audio clients using the ALSA api like cras from ChromiumOS.

Signed-off-by: Girija Kumar Kasinadhuni <gkumar@neverware.com>