x86/kdump: directly find a candidate region when crashkernel=X
I encounter a case where crashkernel=384M, and kaslr is enabled. During the
test, sometimes, the system may fail to reserve region for crash kernel,
although there is much free space above 896MB. It is caused by the
truncation of the candidate region by kaslr kernel. It raises confusion to
the end user that sometimes crashkernel=X works while sometimes fails.
Since on x86, kaslr is a default option, and this corner case is
unavoidable.
This patch simplifies the method suggested in the mail [1]. It just goes
bottom-up to find a candidate region for crashkernel.
There is one trivial thing about the compatibility with old kexec-tools:
if the reserved region is above 896M, then old tool will fail to load
bzImage. But without this patch, the old tool also fail since there is no
memory below 896M can be reserved for crashkernel.

[1]: http://lists.infradead.org/pipermail/kexec/2017-October/019571.html
Signed-off-by: Pingfan Liu <kernelfans@gmail.com>
Cc: Dave Young <dyoung@redhat.com>
Cc: Andrew Morton <akpm@linux-foundation.org>
Cc: Baoquan He <bhe@redhat.com>
Cc: yinghai@kernel.org,
Cc: vgoyal@redhat.com
Cc: kexec@lists.infradead.org