media controller: add properties support
Add support for properties. In this initial implementation properties
can be added to entities and pads. In addition, properties can be
nested.

Most of the changes are straightforward, but I had to make some changes
to the way entities are initialized, since the struct has to be
initialized either when media_entity_pads_init() is called or when
media_device_register_entity() is called, whichever comes first.

The properties list in the entity has to be initialized in both cases,
otherwise you can't add properties to it.

The entity struct is now initialized in media_entity_init and called
from both functions if ent->inited is 0.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>