drm/i915/icl: Fix TypeC legacy HDMI HPD handling
Atm HPD disconnect events on TypeC ports will break things, since we'll
switch the TypeC mode (between Legacy and disconnected modes as well as
among USB DP alternate, Thunderbolt alternate and disconnected modes) on
the fly from the HPD disconnect interrupt work while the port may be
still active.

Even if the port happens to be not active during the disconnect we'd
still have a problem during a subsequent modeset or AUX transfer that
could happen regardless of the port's connected state. For instance the
system resume display mode restore code and userspace could perform a
modeset on the port or userspace could start an AUX transfer even if the
port is in disconnected state.

To fix this keep legacy TypeC ports in TypeC legacy mode whenever we're
not suspended. The legacy mode is a static configuration as opposed to
the Thunderbolt and USB DP alternate modes between which we can switch
dynamically.

We don't have yet an explicit way to detect TypeC legacy ports, but we
can imply that at least in case of HDMI enabled ports, since HDMI can
only be provided in legacy mode. So mark TypeC HDMI enabled ports as
legacy and run the TypeC PHY connect sequence explicitly during driver
loading and system resume. The connect will succeed even if the display
is not connected to begin with (or disappears during the suspended
state) since for legacy ports the PORT_TX_DFLEXDPPMS /
DP_PHY_MODE_STATUS_COMPLETED flag is always set (as opposed to the USB
DP alternate mode where it gets set only when a display is connected).

Correspondingly run the TypeC PHY disconnect sequence during system
suspend and driver unloading, but disconnect during suspend only for
legacy TypeC mode. We will need to disconnect even in USB DP alternate
mode in the future, but atm we don't have a way to reconnect the port
in this mode during resume if the display disappears while being
suspended. So for now punt on this case.

Note that we do not disconnect the port during runtime suspend; in
legacy mode there are no shared HW resources (PHY lanes) with other HW
blocks (USB), so no need to release / reacquire these resources as with
USB DP alternate mode. The only reason to disconnect legacy ports during
system suspend is that the PORT_TX_DFLEXDPPMS /
DP_PHY_MODE_STATUS_COMPLETED flag must be rechecked and the port must be
connected again during system resume. We may also need to turn the check
for this flag into a poll, depending on further tests and clarifications
we are expecting from HW/FW people.

If VBT says that the port provides only DP functionality then we can't
assume that it's a legacy port as for HDMI (since it could as well be
a TBT/DP Alt connector), so we'll solve HPD handling for the DP case
with a different method in the next patch.

Eventually - instead of the above method - we'll depend on an explicit
detection method provided either via a VBT flag or a FW/HW register both
for the HDMI and DP case.

Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=108070
Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=108924
Cc: Paulo Zanoni <paulo.r.zanoni@intel.com>
Cc: Ville Syrjälä <ville.syrjala@linux.intel.com>
Cc: José Roberto de Souza <jose.souza@intel.com>
Cc: Rodrigo Vivi <rodrigo.vivi@intel.com>
Signed-off-by: Imre Deak <imre.deak@intel.com>